requirements.txt
Pillow>=10.0.0
huggingface-hub==0.34.0
peft>=0.10.0
torch>=2.0.0
torchvision>=0.15.0
transformers
git+https://github.com/huggingface/diffusers
